requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
Pillow==10.0.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
psycopg2==2.9.1
